# docs/api-spec.yml
openapi: 3.0.0
info:
  title: Facturas API
  description: |
    REST API for the Facturas multi-tenant invoice management system. Built with Next.js App Router, SQLite (better-sqlite3), and multi-tenant architecture.

    ## Features
    - Multi-tenant invoice management with isolated databases per business
    - Automatic invoice extraction from Gmail attachments (XML/PDF)
    - Confidence scoring and duplicate detection for extracted invoices
    - Full-text search with SQLite FTS5
    - CSV and XLSX export
    - Tags/labels system for invoice categorization
    - API key authentication for public API access
    - Role-based access control (admin/negocio)

    ## Authentication
    - Session-based auth via httpOnly cookies (`session_id`)
    - API key auth for `/api/v1/*` endpoints via `Authorization: Bearer <api_key>`
    - Multi-tenant isolation via `negocio_slug` cookie

  version: 0.1.0
  contact:
    name: Facturas Development Team
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /api/negocios:
    get:
      summary: List negocios
      description: Get all negocios (admin sees all, negocio user sees own)
      tags: [Negocios]
      responses:
        '200':
          description: Negocios list with active slug and user info
    post:
      summary: Create negocio
      description: Create a new business (admin only)
      tags: [Negocios]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                email:
                  type: string
              required: [nombre]
      responses:
        '201':
          description: Negocio created
        '409':
          description: Nombre already exists
  /api/negocios/{slug}:
    patch:
      summary: Update negocio
      description: Update negocio settings (nombre, email, moneda_default). Nombre and email changes limited to once every 6 months.
      tags: [Negocios]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                email:
                  type: string
                moneda_default:
                  type: string
                  default: MXN
      responses:
        '200':
          description: Negocio updated
        '400':
          description: Validation error (e.g. name change too soon)
  /api/negocios/{slug}/select:
    post:
      summary: Select active negocio
      description: Set the active negocio cookie for the session
      tags: [Negocios]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Negocio selected
  /api/facturas:
    get:
      summary: List facturas
      description: |
        Retrieve paginated list of facturas for the active tenant.
        Supports filtering by date range, emisor, estado, moneda, confianza level, revision status, and tags.
        Full-text search via FTS5 when available.
        Currency conversion applied based on negocio default currency.
      tags: [Facturas]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: search
          schema:
            type: string
          description: Full-text search (FTS5) or LIKE fallback
        - in: query
          name: fecha_desde
          schema:
            type: string
            format: date
        - in: query
          name: fecha_hasta
          schema:
            type: string
            format: date
        - in: query
          name: emisor
          schema:
            type: string
          description: Filter by emisor name (LIKE)
        - in: query
          name: estado
          schema:
            type: string
            enum: [pendiente, pagada, cancelada]
        - in: query
          name: moneda
          schema:
            type: string
        - in: query
          name: confianza
          schema:
            type: string
            enum: [alta, media, baja, error]
        - in: query
          name: revision
          schema:
            type: integer
            enum: [0, 1]
        - in: query
          name: etiqueta
          schema:
            type: string
          description: Filter by tag name
      responses:
        '200':
          description: Paginated facturas with currency conversion
          content:
            application/json:
              schema:
                type: object
                properties:
                  facturas:
                    type: array
                    items:
                      $ref: '#/components/schemas/Factura'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: No active tenant
...
  /api/v1/facturas:
    get:
      summary: List facturas (public API)
      description: List facturas via API key authentication. Same filters as internal endpoint.
      tags: [Public API]
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: estado
          schema:
            type: string
            enum: [pendiente, pagada, cancelada]
      responses:
        '200':
          description: Paginated facturas
        '401':
          description: Invalid or missing API key
    post:
      summary: Create factura (public API, write permission required)
      tags: [Public API]
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Factura'
      responses:
        '201':
          description: Factura created
        '401':
          description: Invalid API key or insufficient permissions
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer <api_key>"
  schemas:
    Factura:
      type: object
      properties:
        id:
          type: integer
        emisor_nombre:
          type: string
        emisor_nif:
          type: string
          nullable: true
        emisor_direccion:
          type: string
          nullable: true
        emisor_poblacion:
          type: string
          nullable: true
        emisor_provincia:
          type: string
          nullable: true
        emisor_cp:
          type: string
          nullable: true
        emisor_pais:
          type: string
        emisor_email:
          type: string
          nullable: true
        emisor_telefono:
          type: string
          nullable: true
        receptor_nombre:
          type: string
          nullable: true
        receptor_nif:
          type: string
          nullable: true
        receptor_direccion:
          type: string
          nullable: true
        numero_factura:
          type: string
        fecha_emision:
          type: string
          format: date
        fecha_vencimiento:
          type: string
          format: date
          nullable: true
        tipo_documento:
          type: string
          default: factura
        moneda:
          type: string
          default: MXN
        base_imponible:
          type: number
        tipo_iva:
          type: number
          default: 21.0
        cuota_iva:
          type: number
        total:
          type: number
        descuento:
          type: number
          default: 0
        retencion:
          type: number
          default: 0
        neto:
          type: number
          nullable: true
        metodo_pago:
          type: string
          nullable: true
        estado:
          type: string
          enum: [pendiente, pagada, cancelada]
          default: pendiente
        confianza_score:
          type: number
          description: Confidence score 0.0 to 1.0
        confianza_nivel:
          type: string
          enum: [alta, media, baja, error]
          default: alta
        requiere_revision:
          type: integer
          enum: [0, 1]
          default: 0
        revision_notas:
          type: string
          nullable: true
        total_original:
          type: number
          description: Total in original currency (when currency conversion applied)
        moneda_original:
          type: string
          description: Original currency code
        total_convertido:
          type: number
          description: Total converted to negocio default currency
      required: [emisor_nombre, numero_factura, fecha_emision, base_imponible, cuota_iva, total]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...